<wh> Registrasi dan Login Dashboard
Penambahan fasilitas Back Office untuk pendaftaran dan login, namun demikian untuk akses ke fasilitas lain perlu persetujuan dari Admin

+++ database/migrations/2021_05_17_130901_create_role_menus_table.php
@@ -15,7 +15,7 @@ class CreateRoleMenusTable extends Migration
     {
         Schema::create('role_menus', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->bigInteger('role_id')->unsigned();       
+            $table->integer('role_id')->unsigned();       
             $table->bigInteger('menu_id')->unsigned(); 
             $table->foreign('role_id')
                     ->references('id')->on('roles')
